+++ integrations/aws/utils/misc.py
